Update leetCodePreviewProvider function structure
Move the logic in webviewUtils.ts to leetCodePreviewProvider
Fix missing semicolon
Rename leetcodePreviewProvider to leetCodePreviewProvider
